Move objects to separate game_object folder.
