+++ src/main/java/com/rocketden/main/game_object/Player.java
@@ -1,9 +1,11 @@
-package com.rocketden.main.model;
+package com.rocketden.main.game_object;
 
 import lombok.Getter;
 import lombok.Setter;
 
 import java.util.List;
+
+import com.rocketden.main.model.User;
 
 @Getter
 @Setter
