+++ src/main/java/com/rocketden/main/game_object/Submission.java
@@ -1,4 +1,4 @@
-package com.rocketden.main.model;
+package com.rocketden.main.game_object;
 
 import lombok.Getter;
 import lombok.Setter;
